// src/lib2/models/AxeAnalytiqueDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AxeAnalytiqueDto = {
    id?: string;
    code?: string;
    libelle?: string;
    type?: string;
    responsable?: string;
    parentId?: string;
    typeCentre?: string;
    budgetAnnuel?: number;
    uniteOeuvreCode?: string;
    actif?: boolean;
    compteIds?: Array<string>;
    compteLibelles?: Array<string>;
};